requirements.txt
pytest
wheel
ruamel.yaml.string
sphinx
sphinx-autoapi
furo
Django
ruff
pylint
black
xhtml2pdf